+++ Alarmy-iOS/Alarmy-iOS/Network/Base/RestAPIType.swift
@@ -39,37 +39,3 @@ enum HeaderType {
         }
     }
 }
-
-enum Endpoint {
-    // get 예시
-    case fetch
-    // post 예시
-    case login
-    
-    var restType: RestAPIType {
-        switch self {
-        case .fetch:
-            return .get
-        case .login:
-            return .post
-        }
-    }
-    
-    var url: String {
-        switch self {
-        case .fetch:
-            return ""
-        case .login:
-            return ""
-        }
-    }
-    
-    var header: [String: String] {
-        switch self {
-        case .fetch:
-            HeaderType.none.value
-        case .login:
-            HeaderType.auth.value
-        }
-    }
-}
